Move create cookie method to WebContextHelper (renamed from ContextHelper)
