+++ pac4j-core/src/main/java/org/pac4j/core/context/WebContextHelper.java
@@ -2,7 +2,12 @@ package org.pac4j.core.context;
 
 import org.pac4j.core.util.CommonHelper;
 
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.Collection;
+import java.util.Locale;
 
 /**
  * A helper for the web context.
@@ -10,7 +15,14 @@ import java.util.Collection;
  * @author Jerome Leleu
  * @since 1.8.1
  */
-public final class ContextHelper implements HttpConstants {
+public final class WebContextHelper implements HttpConstants {
+
+    private static ZoneId GMT = ZoneId.of("GMT");
+    /**
+     * Date formats with time zone as specified in the HTTP RFC to use for formatting.
+     * @see <a href="https://tools.ietf.org/html/rfc7231#section-7.1.1.1">Section 7.1.1.1 of RFC 7231</a>
+     */
+    private static DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("EEE, dd MMM yyyy HH:mm:ss zzz", Locale.US).withZone(GMT);
 
     /**
      * Get a specific cookie by its name.
@@ -119,5 +131,51 @@ public final class ContextHelper implements HttpConstants {
      */
     public static boolean isHttps(final WebContext context) {
         return SCHEME_HTTPS.equalsIgnoreCase(context.getScheme());
+    }
+
+    /**
+     * Custom method for adding cookie because the servlet-api version doesn't support SameSite attributes.
+     * @param cookie pac4j Cookie object
+     */
+    public static String createCookieHeader(Cookie cookie) {
+        var builder = new StringBuilder();
+        builder.append(String.format("%s=%s;", cookie.getName(), cookie.getValue()));
+
+        if (cookie.getMaxAge() > -1) {
+            builder.append(String.format(" Max-Age=%s;", cookie.getMaxAge()));
+            long millis = cookie.getMaxAge() > 0 ? System.currentTimeMillis() + (cookie.getMaxAge() * 1000) : 0;
+            Instant instant = Instant.ofEpochMilli(millis);
+            ZonedDateTime time = ZonedDateTime.ofInstant(instant, GMT);
+            builder.append(String.format(" Expires=%s;", DATE_FORMATTER.format(time)));
+        }
+        if (CommonHelper.isNotBlank(cookie.getDomain())) {
+            builder.append(String.format(" Domain=%s;", cookie.getDomain()));
+        }
+        builder.append(String.format(" Path=%s;", CommonHelper.isNotBlank(cookie.getPath()) ? cookie.getPath() : "/"));
+
+        var sameSitePolicy = cookie.getSameSitePolicy() == null ? "none" : cookie.getSameSitePolicy().toLowerCase();
+        switch (sameSitePolicy) {
+            case "strict":
+                builder.append(" SameSite=Strict;");
+                break;
+            case "lax":
+                builder.append(" SameSite=Lax;");
+                break;
+            case "none":
+            default:
+                builder.append(" SameSite=None;");
+                break;
+        }
+        if (cookie.isSecure() || "none".equals(sameSitePolicy)) {
+            builder.append(" Secure;");
+        }
+        if (cookie.isHttpOnly()) {
+            builder.append(" HttpOnly;");
+        }
+        var value = builder.toString();
+        if (value.endsWith(";")) {
+            value = value.substring(0, value.length() - 1);
+        }
+        return value;
     }
 }
