+++ pac4j-core/src/main/java/org/pac4j/core/credentials/extractor/ParameterExtractor.java
@@ -1,6 +1,6 @@
 package org.pac4j.core.credentials.extractor;
 
-import org.pac4j.core.context.ContextHelper;
+import org.pac4j.core.context.WebContextHelper;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.context.session.SessionStore;
 import org.pac4j.core.credentials.Credentials;
@@ -37,9 +37,9 @@ public class ParameterExtractor implements CredentialsExtractor {
 
     @Override
     public Optional<Credentials> extract(final WebContext context, final SessionStore sessionStore) {
-        if (ContextHelper.isGet(context) && !supportGetRequest) {
+        if (WebContextHelper.isGet(context) && !supportGetRequest) {
             throw new CredentialsException("GET requests not supported");
-        } else if (ContextHelper.isPost(context) && !supportPostRequest) {
+        } else if (WebContextHelper.isPost(context) && !supportPostRequest) {
             throw new CredentialsException("POST requests not supported");
         }
