+++ pac4j-core/src/main/java/org/pac4j/core/matching/matcher/StrictTransportSecurityMatcher.java
@@ -1,6 +1,6 @@
 package org.pac4j.core.matching.matcher;
 
-import org.pac4j.core.context.ContextHelper;
+import org.pac4j.core.context.WebContextHelper;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.context.session.SessionStore;
 import org.pac4j.core.util.CommonHelper;
@@ -28,7 +28,7 @@ public class StrictTransportSecurityMatcher implements Matcher {
 
     @Override
     public boolean matches(final WebContext context, final SessionStore sessionStore) {
-        if (ContextHelper.isHttpsOrSecure(context)) {
+        if (WebContextHelper.isHttpsOrSecure(context)) {
             context.setResponseHeader("Strict-Transport-Security", "max-age=" + maxAge + " ; includeSubDomains");
         }
         return true;
